1️⃣ 코드 구조화 작업 중
